Changed `options_pixie_process_bulk_action` action to be `options_pixie_process_action` filter.
This allows the same process flow to be used for row level actions.
First param is now whether a redirect is required, true/false,
default is false.
Second param is action name.
Third is single or array of `option_id`s.
Fourth is `blog_id`, blank if single site install.